personal notes finished mvp update

+++ App.js
@@ -3,9 +3,10 @@ import {
   View,
   Text,
   StyleSheet,
-  Modal,
+  ActivityIndicator,
   TextInput,
   TouchableOpacity,
+  Alert,
 } from 'react-native';
 import {heightPercentageToDP as hp} from 'react-native-responsive-screen';
 import MasonryList from '@react-native-seoul/masonry-list';
@@ -21,10 +22,12 @@ import {
 } from '@react-native-firebase/firestore';
 import FAIcon from 'react-native-vector-icons/FontAwesome';
 import FAIcon5 from 'react-native-vector-icons/FontAwesome5';
+import SIMPLEIcon from 'react-native-vector-icons/SimpleLineIcons';
 
 export default function App() {
   const [notes, setNotes] = useState([]);
   const [selectedNote, setSelectedNote] = useState(null);
+  const [isNoteLoading, setIsNoteLoading] = useState(false);
 
   console.log('noes', notes);
 
@@ -62,8 +65,8 @@ export default function App() {
       id: docRef.id,
     };
 
-    setNotes(prev => [...prev, newNote]);
     setSelectedNote(newNote);
+    setIsNoteLoading(false);
   };
 
   const updateNote = async (updateNoteTitle, updateNoteContent) => {
@@ -85,16 +88,28 @@ export default function App() {
     setNotes(prev =>
       prev.map(note => (note.id === selectedNote.id ? updatedNote : note)),
     );
-
     setSelectedNote(updatedNote);
+    setIsNoteLoading(false);
   };
 
   const deleteNote = async () => {
     const db = getFirestore();
     const noteRef = doc(db, 'Notes', selectedNote.id);
 
-    await deleteDoc(noteRef);
-    setSelectedNote(null);
+    Alert.alert('Confirm', 'Are you sure to delete note?', [
+      {
+        text: 'OK',
+        onPress: async () => {
+          await deleteDoc(noteRef);
+          setSelectedNote(null);
+        },
+      },
+      {
+        text: 'Cancel',
+        onPress: () => console.log('Cancel Pressed'),
+        style: 'cancel',
+      },
+    ]);
   };
 
   const Card = ({title, content, id}) => (
@@ -118,11 +133,21 @@ export default function App() {
     };
 
     const addOrUpdate = () => {
+      setIsNoteLoading(true);
       if (newNotes.id.trim()) {
         updateNote(newNotes.title, newNotes.content);
       } else {
         addNewNote(newNotes.title, newNotes.content);
       }
+    };
+
+    const parsedContent = text => {
+      const parsedText = text.replace(
+        /- (\w)/g,
+        (match, p1) => `\u25CF ${p1.toUpperCase()}`,
+      );
+      console.log('parsed text', parsedText);
+      return parsedText;
     };
 
     return (
@@ -132,37 +157,59 @@ export default function App() {
             flexDirection: 'row',
             alignItems: 'center',
             justifyContent: 'space-between',
+            marginBottom: hp('1%'),
           }}>
           <TouchableOpacity
             onPress={removeSelectedNote}
-            style={{padding: hp('1%')}}>
+            style={{padding: hp('1%'), paddingVertical: hp('2%')}}>
             <FAIcon5 name="chevron-left" size={hp('3%')} color="#595550" />
           </TouchableOpacity>
           <View>
-            <TextInput
-              placeholder="Note Title"
-              style={{fontSize: hp('3%')}}
-              value={newNotes.title}
-              onChangeText={text =>
-                setNewNotes(prev => ({...prev, title: text}))
-              }
-            />
+            {!isNoteLoading && (
+              <TextInput
+                placeholder="Note Title"
+                placeholderTextColor="#ababab"
+                style={{fontSize: hp('2.75%')}}
+                value={newNotes.title}
+                onChangeText={text =>
+                  setNewNotes(prev => ({...prev, title: text}))
+                }
+              />
+            )}
           </View>
-          <TouchableOpacity style={{padding: hp('1%')}} onPress={deleteNote}>
-            <FAIcon name="trash-o" size={hp('3.5%')} color="red" />
+          <TouchableOpacity
+            style={{padding: hp('1%'), paddingVertical: hp('2%')}}
+            onPress={selectedNote.id.trim() ? deleteNote : null}
+            disabled={!selectedNote.id.trim()}
+            accessibilityLabel="Delete note">
+            <FAIcon
+              name="trash-o"
+              size={hp('3.5%')}
+              color={selectedNote.id.trim() ? 'red' : 'transparent'}
+            />
           </TouchableOpacity>
         </View>
 
-        <View style={{flex: 1}}>
+        <View style={{flex: 1, gap: hp('2%')}}>
           <View style={{flex: 1}}>
-            <TextInput
-              placeholder="Note content..."
-              value={newNotes.content}
-              onChangeText={text =>
-                setNewNotes(prev => ({...prev, content: text}))
-              }
-              multiline
-            />
+            {isNoteLoading ? (
+              <ActivityIndicator color="#595550" size="large" />
+            ) : (
+              <TextInput
+                placeholder="Note content..."
+                placeholderTextColor="#ababab"
+                value={parsedContent(newNotes.content)}
+                onChangeText={text =>
+                  setNewNotes(prev => ({...prev, content: text}))
+                }
+                style={{
+                  height: '100%',
+                  textAlignVertical: 'top',
+                  padding: hp('1%'),
+                }}
+                multiline
+              />
+            )}
           </View>
           <TouchableOpacity style={styles.AddNewNoteBtn} onPress={addOrUpdate}>
             <Text
@@ -183,7 +230,22 @@ export default function App() {
   const HomeScreen = () => {
     return (
       <View style={styles.container}>
+        <View
+          style={{
+            flexDirection: 'row',
+            justifyContent: 'center',
+            alignItems: 'center',
+            gap: hp('2%'),
+            marginBottom: hp('2%'),
+          }}>
+          <Text
+            style={{color: '#373533', fontWeight: 'bold', fontSize: hp('3%')}}>
+            UtopicUnicorn's notes
+          </Text>
+          <SIMPLEIcon name="notebook" size={hp('3%')} color="#595550" />
+        </View>
         <MasonryList
+          refreshControl={false}
           data={notes}
           style={{gap: hp('2%')}}
           keyExtractor={(item, index) => item.id || index.toString()}
